Rating finally working

+++ app/src/main/java/it/polito/timebanking/ui/messages/MessageListFragment.kt
@@ -157,19 +157,19 @@ class MessageListFragment : Fragment() {
             dialog.setView(dialogView)
 
             dialog.setPositiveButton("Confirm") { _, _ ->
-                val rating = dialogView.findViewById<RatingBar>(R.id.ratingBar).rating
+                val rating = dialogView.findViewById<RatingBar>(R.id.ratingBar).rating.toDouble()
                 val comment = dialogView.findViewById<EditText>(R.id.comment).text.toString()
                 if (userIsProducer) {
-                    val rate = RateData(rating.toDouble(), comment, timeslot.title, job.userProducerID, job.userConsumerID, false)
+                    val rate = RateData(rating, comment, timeslot.title, job.userProducerID, job.userConsumerID, false)
                     FirebaseFirestore.getInstance().collection("ratings").add(rate).addOnSuccessListener {
-                        FirebaseFirestore.getInstance().collection("users").document(job.userConsumerID).update("jobsRated", FieldValue.increment(1), "score", FieldValue.increment(rating.toLong()))
+                        FirebaseFirestore.getInstance().collection("users").document(job.userConsumerID).update("jobsRated", FieldValue.increment(1), "score", FieldValue.increment(rating))
                         Snackbar.make(binding.root, "Rated successfully", Snackbar.LENGTH_SHORT).show()
                     }
                 }
                 else {
-                    val rate = RateData(rating.toDouble(), comment, timeslot.title, job.userConsumerID, job.userProducerID, true)
+                    val rate = RateData(rating, comment, timeslot.title, job.userConsumerID, job.userProducerID, true)
                     FirebaseFirestore.getInstance().collection("ratings").add(rate).addOnSuccessListener {
-                        FirebaseFirestore.getInstance().collection("users").document(job.userProducerID).update("jobsRated", FieldValue.increment(1), "score", FieldValue.increment(rating.toLong()))
+                        FirebaseFirestore.getInstance().collection("users").document(job.userProducerID).update("jobsRated", FieldValue.increment(1), "score", FieldValue.increment(rating))
                         Snackbar.make(binding.root, "Rated successfully", Snackbar.LENGTH_SHORT).show()
                     }
                 }
